Fix #353 - support exception filter, and avoid NestJS bug

+++ test/executable/setup.js
@@ -17,9 +17,17 @@ const main = async () => {
             content.devDependencies,
         ])
             for (const [name, version] of Object.entries(dependencies ?? {})) {
-                if (name.includes("@nestia")) continue;
+                if (name.includes("@nestia") || name.includes("@nestjs")) continue;
                 devDependencies.push([name, version]);
             }
+        process.chdir(directory(modulo));
+        cp.execSync("npm install", { stdio: "ignore" });
+        cp.execSync("npm run build", { stdio: "ignore" });
+        cp.execSync("npm pack", { stdio: "ignore" });
+        devDependencies.push([
+            content.name, 
+            `../packages/${modulo}/${content.name.replace("@", "").replace("/", "-")}-${content.version}.tgz`
+        ]);
     }
     devDependencies.sort((a, b) => a[0].localeCompare(b[0]));
 
@@ -36,8 +44,10 @@ const main = async () => {
         "utf8",
     );
 
-    await fs.promises.rm(`${__dirname}/../node_modules`, { recursive: true });
-    await fs.promises.rm(`${__dirname}/../package-lock.json`);
+    if (fs.existsSync(`${__dirname}/../node_modules`))
+        await fs.promises.rm(`${__dirname}/../node_modules`, { recursive: true });
+    if (fs.existsSync(`${__dirname}/../package-lock.json`))
+        await fs.promises.rm(`${__dirname}/../package-lock.json`);
     cp.execSync("npm install", { cwd: `${__dirname}/..`, stdio: "inherit" });
 };
